// query-node/types/generated/1001/types-lookup.ts
export interface PalletWorkingGroupStakePolicy extends Struct {
  readonly stakeAmount: u128;
  readonly leavingUnstakingPeriod: u32;
}
// ...
export interface PalletWorkingGroupOpeningType extends Enum {
  readonly isLeader: boolean;
  readonly isRegular: boolean;
  readonly type: "Leader" | "Regular";
}

/** @name PalletWorkingGroupApplyOnOpeningParams (217) */
export interface PalletWorkingGroupApplyOnOpeningParams extends Struct {
  readonly memberId: u64;
  readonly openingId: u64;
  readonly roleAccountId: AccountId32;
  readonly rewardAccountId: AccountId32;
  readonly description: Bytes;
  readonly stakeParameters: PalletWorkingGroupStakeParameters;
}

/** @name PalletWorkingGroupStakeParameters (218) */
export interface PalletWorkingGroupStakeParameters extends Struct {
  readonly stake: u128;
  readonly stakingAccountId: AccountId32;
}

/** @name PalletWorkingGroupInstance1 (219) */
export type PalletWorkingGroupInstance1 = Null;

/** @name PalletWorkingGroupRewardPaymentType (220) */
export interface PalletWorkingGroupRewardPaymentType extends Enum {
  readonly isMissedReward: boolean;
  readonly isRegularReward: boolean;
  readonly type: "MissedReward" | "RegularReward";
}